# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGI Judd's Idea Catalog API",
    "description": "Transform LLM conversations into actionable Linear projects",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/llm-defaults/{prompt_type}": {
      "get": {
        "summary": "Get Llm Defaults",
        "description": "Get the current default LLM settings for a specific prompt type.",
        "operationId": "get_llm_defaults_api_llm_defaults__prompt_type__get",
        "parameters": [
          {
            "name": "prompt_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LLMDefaultsResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Get Llm Defaults Api Llm Defaults  Prompt Type  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Llm Defaults",
        "description": "Update the default LLM settings for a specific prompt type.",
        "operationId": "update_llm_defaults_api_llm_defaults__prompt_type__put",
        "parameters": [
          {
            "name": "prompt_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Type"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMDefaultsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LLMDefaultsUpdateResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Update Llm Defaults Api Llm Defaults  Prompt Type  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/llm-prompts/{prompt_type}": {
      "get": {
        "summary": "Get Llm Prompt",
        "description": "Get the active LLM prompt for a given type.\nReturns the prompt from database if available, otherwise indicates default is being used.",
        "operationId": "get_llm_prompt_api_llm_prompts__prompt_type__get",
        "parameters": [
          {
            "name": "prompt_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LLMPromptResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Get Llm Prompt Api Llm Prompts  Prompt Type  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Llm Prompt",
        "description": "Create or update an LLM prompt for a given type.\nThis will deactivate any existing active prompt of the same type.",
        "operationId": "create_llm_prompt_api_llm_prompts__prompt_type__post",
        "parameters": [
          {
            "name": "prompt_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Type"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMPromptCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LLMPromptResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Create Llm Prompt Api Llm Prompts  Prompt Type  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Llm Prompt",
        "description": "Deactivate the active LLM prompt for a given type.\nThis will revert to using the default prompt from code.",
        "operationId": "delete_llm_prompt_api_llm_prompts__prompt_type__delete",
        "parameters": [
          {
            "name": "prompt_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LLMPromptDeleteResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Delete Llm Prompt Api Llm Prompts  Prompt Type  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/llm-prompts/{prompt_type}/default": {
      "get": {
        "summary": "Get Default Llm Prompt",
        "description": "Get the default LLM prompt for a given type.\nThis always returns the hardcoded default prompt, ignoring any custom prompts.",
        "operationId": "get_default_llm_prompt_api_llm_prompts__prompt_type__default_get",
        "parameters": [
          {
            "name": "prompt_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LLMPromptResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response Get Default Llm Prompt Api Llm Prompts  Prompt Type  Default Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response schema."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMDefault": {
        "properties": {
          "llm_provider": {
            "type": "string",
            "minLength": 1,
            "title": "Llm Provider",
            "description": "LLM provider name"
          },
          "llm_model": {
            "type": "string",
            "minLength": 1,
            "title": "Llm Model",
            "description": "LLM model name"
          }
        },
        "type": "object",
        "required": [
          "llm_provider",
          "llm_model"
        ],
        "title": "LLMDefault",
        "description": "Model for LLM default settings."
      },
      "LLMDefaultsResponse": {
        "properties": {
          "current_default": {
            "$ref": "#/components/schemas/LLMDefault",
            "description": "Current default LLM settings"
          }
        },
        "type": "object",
        "required": [
          "current_default"
        ],
        "title": "LLMDefaultsResponse",
        "description": "Response model for getting LLM defaults."
      },
      "LLMDefaultsUpdateRequest": {
        "properties": {
          "llm_provider": {
            "type": "string",
            "minLength": 1,
            "title": "Llm Provider",
            "description": "LLM provider name"
          },
          "llm_model": {
            "type": "string",
            "minLength": 1,
            "title": "Llm Model",
            "description": "LLM model name"
          }
        },
        "type": "object",
        "required": [
          "llm_provider",
          "llm_model"
        ],
        "title": "LLMDefaultsUpdateRequest",
        "description": "Request model for updating LLM defaults."
      },
      "LLMDefaultsUpdateResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Success message"
          },
          "updated_default": {
            "$ref": "#/components/schemas/LLMDefault",
            "description": "The updated default settings"
          }
        },
        "type": "object",
        "required": [
          "message",
          "updated_default"
        ],
        "title": "LLMDefaultsUpdateResponse",
        "description": "Response model for updating LLM defaults."
      },
...
      "LLMPromptCreateRequest": {
        "properties": {
          "prompt_type": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt Type",
            "description": "Type of prompt"
          },
          "system_prompt": {
            "type": "string",
            "minLength": 1,
            "title": "System Prompt",
            "description": "The system prompt text"
          }
        },
        "type": "object",
        "required": [
          "prompt_type",
          "system_prompt"
        ],
        "title": "LLMPromptCreateRequest",
        "description": "Request model for creating a new LLM prompt."
      },
      "LLMPromptDeleteResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Success or error message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "LLMPromptDeleteResponse",
        "description": "Response model for LLM prompt deletion."
      },
      "LLMPromptResponse": {
        "properties": {
          "system_prompt": {
            "type": "string",
            "title": "System Prompt",
            "description": "The system prompt text"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "description": "Whether this is using the default prompt from code"
          }
        },
        "type": "object",
        "required": [
          "system_prompt",
          "is_default"
        ],
        "title": "LLMPromptResponse",
        "description": "Response model for LLM prompt operations."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}